fix: store and show package durations as raw hours (#33)
Package durations (duration.minHours/maxHours, durationOfSetup,
durationPerPerson) are stored as raw hours everywhere, decimals allowed
(1h30m = 1.5). Customer pages no longer treat them as minutes; they now
format as "1 hr 30 min" / "12 hrs" via a shared formatHours helper,
which the vendor publish summaries also use. Vendor flows save with
parseFloat so fractional hours are not truncated on re-save. Removes the
now-unused formatMinutes helper.

+++ src/features/packages/flows/makeup/MakeupFlow.tsx
@@ -502,8 +502,8 @@ export default function MakeupFlow({ onExitFlow }: { onExitFlow?: () => void }) 
                         ].join(', ')
                     },
                     // Makeup Artist specific fields
-                    durationPerPerson: parseInt(teamDurationPerPerson) || 0,
-                    durationOfSetup: parseInt(teamDurationOfSetup) || 0,
+                    durationPerPerson: parseFloat(teamDurationPerPerson) || 0,
+                    durationOfSetup: parseFloat(teamDurationOfSetup) || 0,
                     trialOffered: trialOffered === 'Yes',
                     parallelServicingPossible: parallelServicing === 'Yes',
                     crewSize: {
